Add files via upload
In this Fuel consumption prediction Project based on the some Machine Learning Algorithm(i.e -linear regression). Here i created one frontend web page with the help of basic html and css on that page we can take the input from user's engine size and it will predict the Fuel consumption . In this project i work on the flask framework for connectivity of frontend and backend and use python in backend and also imported some libraries (i.e -pandas , numpy , pickle , matplotlib , seaborn).